fix(db): widen count columns to BIGINT to stop 32-bit overflow
The scan-tick bulk write into `tickers` was failing every cycle with
psycopg NumericValueOutOfRange: a high-turnover name (ADTX ~5.28B shares)
exceeds the 32-bit INTEGER max, and one overflowing row aborted the entire
batch — so scores stopped persisting AND Sentry got flooded with thousands
of identical DataErrors (exhausting the 5k/period error quota).

- tickers.volume INTEGER -> BIGINT (the active fire)
- institutional_holdings.shares -> BIGINT (a fund can hold >2.1B shares)
- ipo_events.shares_offered -> BIGINT (latent same-class overflow)
- migration 0034 (Postgres-only; SQLite INTEGER is already 64-bit)
- regression test asserting all three stay BigInteger

+++ backend/app/models/calendar_events.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 
 from datetime import date, datetime
 
-from sqlalchemy import Date, DateTime, Float, Integer, String, func
+from sqlalchemy import BigInteger, Date, DateTime, Float, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db import Base
@@ -20,7 +20,7 @@ class IPOEvent(Base):
     expected_date: Mapped[date] = mapped_column(Date, nullable=False, index=True)
     price_low: Mapped[float | None] = mapped_column(Float, nullable=True)
     price_high: Mapped[float | None] = mapped_column(Float, nullable=True)
-    shares_offered: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    shares_offered: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     status: Mapped[str] = mapped_column(String(20), default="upcoming", nullable=False)
     # upcoming | priced | trading | postponed | withdrawn
     lead_underwriter: Mapped[str | None] = mapped_column(String(120), nullable=True)
